add a working test and clear up the clutter and start writing doc.go
