+++ plugins/manager/doc.go
@@ -1,0 +1,2 @@
+// Package manager provides functionality for
+package manager
